Fix SA1139 (Use literal suffix notation instead of casting).

+++ Samples/Hast.Samples.Consumer/SampleRunners/Posit32CalculatorSampleRunner.cs
@@ -28,7 +28,7 @@ namespace Hast.Samples.Consumer.SampleRunners
                 hastlayer,
                 hardwareRepresentation.HardwareGenerationConfiguration);
 
-            positCalculator.CalculatePowerOfReal(100000, (float)1.0001, hastlayer, hardwareRepresentation.HardwareGenerationConfiguration);
+            positCalculator.CalculatePowerOfReal(100000, 1.0001F, hastlayer, hardwareRepresentation.HardwareGenerationConfiguration);
 
             var numbers = new int[Posit32Calculator.MaxDegreeOfParallelism];
             for (int i = 0; i < Posit32Calculator.MaxDegreeOfParallelism; i++)
@@ -45,8 +45,8 @@ namespace Hast.Samples.Consumer.SampleRunners
 
             for (var i = 0; i < 100000; i++)
             {
-                if (i % 2 == 0) posit32Array[i] = new Posit32((float)0.25 * 2 * i).PositBits;
-                else posit32Array[i] = new Posit32((float)0.25 * -2 * i).PositBits;
+                if (i % 2 == 0) posit32Array[i] = new Posit32(0.25F * 2 * i).PositBits;
+                else posit32Array[i] = new Posit32(0.25F * -2 * i).PositBits;
             }
 
             var positsInArraySum = positCalculator.AddPositsInArray(
@@ -70,9 +70,9 @@ namespace Hast.Samples.Consumer.SampleRunners
 
             Console.WriteLine();
 
-            positCalculator.CalculatePowerOfReal(100000, (float)1.0001);
+            positCalculator.CalculatePowerOfReal(100000, 1.0001F);
             sw = Stopwatch.StartNew();
-            var powerOfReal = positCalculator.CalculatePowerOfReal(100000, (float)1.0001);
+            var powerOfReal = positCalculator.CalculatePowerOfReal(100000, 1.0001F);
             sw.Stop();
 
             Console.WriteLine("Result of power of real number: " + powerOfReal);
@@ -100,8 +100,8 @@ namespace Hast.Samples.Consumer.SampleRunners
 
             for (var i = 0; i < 100000; i++)
             {
-                if (i % 2 == 0) posit32Array[i] = new Posit32((float)0.25 * 2 * i).PositBits;
-                else posit32Array[i] = new Posit32((float)0.25 * -2 * i).PositBits;
+                if (i % 2 == 0) posit32Array[i] = new Posit32(0.25F * 2 * i).PositBits;
+                else posit32Array[i] = new Posit32(0.25F * -2 * i).PositBits;
             }
 
             positCalculator.AddPositsInArray(posit32Array);
